+++ HISTORY.md
@@ -4,6 +4,7 @@
 * Added floating Save Changes button on the settings page
 * Added support for multicurrency stores
 * Store configuration is now preserved if Test Setup fails
+* Plugin activation via WP-CLI supported
 
 = 3.9.1  =
 * Added testmode support for simulated test payments via Blockonomics dashboard